made changes in line with reviewer's request
